requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
SQLAlchemy
psycopg2-binary
python-dotenv